requirements.txt
pymongo~=3.11.4
DateTime~=4.9
python-dateutil~=2.8.2
networkx~=2.8.8
nltk~=3.8.1
PyYAML~=5.3.1
schedule~=0.6.0
tqdm~=4.50.2
numpy~=1.24.1
scikit-learn~=1.2.2
matplotlib~=3.6.3
seaborn~=0.12.2
click~=7.1.2
igraph~=0.10.4
cdlib~=0.2.6
graphviz~=0.20.1
gurobipy~=10.0.1
bayanpy~=0.7.4
pygraphviz~=1.11
atproto~=0.0.46